Move grid algorithms into a separate file
